Added Dungeon Game
This would have been a stellar opportunity to use git for some awesome
documentation practice, but I got too carried away... :/ sorry folks...

I'm mostly done here and I suspect since the next project is to do the
same thing but using Classes and Types (?) then I'll not be returning to
this. It was a blast though!